refactor: extract V3 homepage adapters and flatten card context
Moves per-card data shaping out of HomepageView.get_v3_context_data
and into ak/homepage.py:
- build_community_posts, build_join_developers_links,
  build_get_started_code, build_library_intro, hero_image_context
- WHY_BOOST_CARDS, HELLO_WORLD_SNIPPET, HERO_* constants at module
  scope so they allocate once at import, not per request

Flattens the card context wrappers ({'items': [...]}/'bars': [...])
so each context key carries a flat list. Chrome (heading, CTA,
variant, theme) moves into the homepage template, matching how
other V3 surfaces already pass these literals.

Co-locates event_to_card_dict + upcoming_events next to the rest of
the calendar pipeline in core/calendar.py.

HomepageView.get_v3_context_data goes from ~150 lines to ~40 lines
of one-key-per-line orchestration.
