+++ core/calendar.py
@@ -90,3 +90,29 @@ def events_by_month(events):
         events_by_month[month_year].append(event)
 
     return events_by_month
+
+
+def event_to_card_dict(event):
+    """Shape a calendar event (start, end, name, description) for the event card.
+
+    Passes the start as a date object; the template formats it.
+    """
+    return {
+        "title": event.get("name"),
+        "description": event.get("description"),
+        "date": event["start"],
+    }
+
+
+def upcoming_events(events_by_month, limit=4):
+    """Flatten the month-keyed dict from HomepageView.get_events() into the
+    next `limit` events, soonest first.
+    """
+    events = [
+        event
+        for month_events in (events_by_month or {}).values()
+        for event in month_events
+        if event.get("start")
+    ]
+    events.sort(key=lambda event: event["start"])
+    return [event_to_card_dict(event) for event in events[:limit]]
