+++ templates/v3/homepage.html
@@ -28,18 +28,14 @@
         {% include "v3/includes/_event_card.html" with heading="Upcoming Events" items=upcoming_events variant="list" empty_message="No Upcoming Events" primary_cta_label="Download latest release" primary_cta_url=release_url secondary_cta_label="View events calendar" secondary_cta_url=calendar_url %}
       </div>
       <div class="item-c responsive-limit-2-items">
-        {% with data=posts_from_the_boost_community %}
-          {% include "v3/includes/_post_card.html" with heading=data.heading items=data.items variant=data.variant theme=data.theme primary_cta_label=data.primary_cta_label primary_cta_url=posts_url %}
-        {% endwith %}
+        {% include "v3/includes/_post_card.html" with heading="Posts from the Boost Community" items=community_posts variant="card" theme="teal" primary_cta_label="View all posts" primary_cta_url=posts_url %}
       </div>
         <div class="item-d">
           {% url 'library-detail' version_slug='latest' library_slug=get_started_code.library_slug as get_started_url %}
           {% include "v3/includes/_code_block_card.html" with heading=get_started_code.heading code=get_started_code.code language=get_started_code.language block_variant="grey-bg" button_text="View library" button_url=get_started_url button_aria_label="Explore examples" %}
         </div>
       <div class="item-e homepage-join-card">
-        {% with data=join_developers_building_the_future_of_cpp %}
-          {% include "v3/includes/_join_card.html" with heading="Join developers building the future of C++" items=data.items variant="card" theme="yellow" primary_cta_label="Explore the community" primary_cta_url=community_url %}
-        {% endwith %}
+        {% include "v3/includes/_join_card.html" with heading="Join developers building the future of C++" items=join_developers_links variant="card" theme="yellow" primary_cta_label="Explore the community" primary_cta_url=community_url %}
       </div>
       <div class="item-f">
         {% include "v3/includes/_install_card.html" with title=install_card_title|default:"Install Boost and get started in your terminal." card_id=install_card_id|default:"hero-home-install" %}
@@ -50,9 +46,7 @@
     </div>
     <div class="card-masonry-bottom py-large">
       <div class="item-a responsive-limit-5-items">
-        {% with data=stats_in_numbers %}
-          {% include "v3/includes/_stats_in_numbers.html" with heading="Boost in numbers" description="Every bar tells a story; the collective effort of a community building Boost" bars=data.bars theme="default" primary_cta_label="Boost Org on GitHub" primary_cta_url="https://github.com/boostorg/boost" %}
-        {% endwith %}
+        {% include "v3/includes/_stats_in_numbers.html" with heading="Boost in numbers" description="Every bar tells a story; the collective effort of a community building Boost" bars=commits_data theme="default" primary_cta_label="Boost Org on GitHub" primary_cta_url="https://github.com/boostorg/boost" %}
       </div>
       <div class="item-b">
         {% with data=library_intro %}
@@ -60,9 +54,7 @@
         {% endwith %}
       </div>
       <div class="item-c">
-        {% with data=testimonial_data %}
-          {% include "v3/includes/_testimonial_card.html" with heading="What engineers are saying" testimonials=data.testimonials %}
-        {% endwith %}
+        {% include "v3/includes/_testimonial_card.html" with heading="What engineers are saying" testimonials=testimonial_cards %}
       </div>
       <div class="item-d license-card">
         {% large_static 'img/v3/solo-images/capy-hot-tub.png' as license_card_image %}
